- Change sortOnDisk option to sortInMemory - Fix horrible cleaner bug - Trivial optimizations to cleaner code - more significant ones coming soon.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2850 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/gatk/walkers/indels/IntervalCleanerWalker.java
@@ -715,7 +715,7 @@ public class  IntervalCleanerWalker extends LocusWindowWalker<Integer, Integer> 
             if ( originalMismatchBases[i] > totalOriginalBases[i] * MISMATCH_THRESHOLD )  {
                 didMismatch = true;
                 originalMismatchColumns++;
-                if ( (cleanedMismatchBases[i] / totalCleanedBases[i]) > (originalMismatchBases[i] / totalOriginalBases[i]) * (1.0 - MISMATCH_COLUMN_CLEANED_FRACTION) ) {
+                if ( ((double)cleanedMismatchBases[i] / (double)totalCleanedBases[i]) > ((double)originalMismatchBases[i] / (double)totalOriginalBases[i]) * (1.0 - MISMATCH_COLUMN_CLEANED_FRACTION) ) {
                     stillMismatches = true;
                     cleanedMismatchColumns++;
                 }
